Edit stylesheet
Moved declerations from class selector .main
to <html> element as a type selector for cleaner
code. This also fixed the breaks I had in background
color as I had desired to have the entire page
this tan type color. The .center class selector
is commented out because this solution didn't work
as it stretched out the width of the image.